feat : 활동상태에 따라 멤버 조회하여 반환하도록 service 구현
