+++ BE/eeos/src/main/java/com/blackcompany/eeos/member/application/usecase/GetMembersByActiveStatus.java
@@ -1,0 +1,7 @@
+package com.blackcompany.eeos.member.application.usecase;
+
+import com.blackcompany.eeos.member.application.dto.QueryMemberResponse;
+
+public interface GetMembersByActiveStatus {
+	QueryMemberResponse execute(String activeStatus);
+}
